Major Push
-Added New Library to installation to handle the api encryption
-deleted a lot of unnecessary html files. They were leftovers from the
template that I was keeping for reference, but I've out grown the need
for them.
-Completed Auth endpoint: authenticates the user via the pearson apis,
and stores all the necessary info such as access token and refresh token
-completed the user endpoint: retrieves user info if the user is logged
in
-completed the course endpoint: retrieves courses that the user is
enrolled in, as well as the courses' respective instructor and teachers
assistant
-created template pages for the following: chat, clique view, course
view
-added a bunch of containers on home and the previously mentioned pages.
This will serve as a example of what the UI is expected to look like.